fix bug

+++ src/draw_raycast.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   draw_raycast.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: raperez- <raperez-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: raperez- <raperez-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/07 12:25:19 by ozamora-          #+#    #+#             */
-/*   Updated: 2025/06/13 17:03:50 by raperez-         ###   ########.fr       */
+/*   Updated: 2025/06/13 21:40:15 by raperez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ void	draw_ray(mlx_image_t *img, t_ray ray, int mult)
 	}
 }
 
-int	aux_pe(t_game *game, mlx_texture_t **texture, t_ray ray, t_point *txt)
+double	aux_pe(t_game *game, mlx_texture_t **texture, t_ray ray, t_point *txt)
 {
 	*texture = get_texture(game, ray);
 	if (!texture || !*texture)
